redesign tstate functionality to allow for parallel exec
